R: split up LoggerTests into several files
