+++ Pocket.Logger/Tests/LoggerTests.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using FluentAssertions;
 using System.Linq;
-using System.Threading.Tasks;
 using Xunit;
 using Xunit.Abstractions;
 using static Pocket.Log;
@@ -31,7 +30,7 @@ namespace Pocket.Tests
         {
             var log = new List<LogEntry>();
 
-            using (Log.Subscribe(log.Add))
+            using (Subscribe(log.Add))
             {
                 Info("hello");
             }
@@ -201,233 +200,6 @@ namespace Pocket.Tests
         }
 
         [Fact]
-        public void Log_Section_can_log_on_enter()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnEnterAndExit().Info("starting..."))
-            {
-                log.Should().ContainSingle(e => e.Message.Contains("starting..."));
-            }
-        }
-
-        [Fact]
-        public void Log_OnExit_logs_a_Stop_event_on_exit()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnExit())
-            {
-            }
-
-            log.Should().HaveCount(1);
-            log.Last().IsSectionComplete.Should().BeTrue();
-            log.Last().IsSectionSuccessful.Should().BeTrue();
-        }
-
-        [Fact]
-        public void Log_OnEnterAndExit_logs_a_Start_event_on_enter_and_a_Stop_event_on_exit()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnEnterAndExit())
-            {
-            }
-
-            log.Should().HaveCount(2);
-            log[0].IsSectionComplete.Should().BeFalse();
-            log[1].IsSectionSuccessful.Should().BeTrue();
-        }
-
-        [Fact]
-        public void Log_entries_within_a_section_share_an_id_when_specified()
-        {
-            var log = new List<LogEntry>();
-            var sectionId ="the-section-id";
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit(id: sectionId))
-            {
-                section.Info("hello");
-            }
-
-            log.Select(e => e.SectionId).Should().OnlyContain(id => id == sectionId);
-        }
-
-        [Fact]
-        public void Log_entries_within_a_section_share_an_id_when_not_specified()
-        {
-            var log = new List<LogEntry>();
-            var sectionId = Guid.NewGuid().ToString();
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit(id: sectionId))
-            {
-                section.Info("hello");
-            }
-
-            log.Select(e => e.SectionId).Distinct().Should().HaveCount(1);
-        }
-
-        [Fact]
-        public async Task Log_Section_records_timing_when_completed()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnExit())
-            {
-                await Task.Delay(500);
-            }
-
-            log.Last()
-               .ElapsedMilliseconds
-               .Should()
-               .BeGreaterOrEqualTo(500);
-        }
-
-        [Fact]
-        public async Task Log_Section_records_timings_for_checkpoints()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnEnterAndExit())
-            {
-                await Task.Delay(200);
-            }
-
-            log[0].ElapsedMilliseconds.Should().BeInRange(0, 50);
-            log[1].ElapsedMilliseconds.Should().BeGreaterOrEqualTo(200);
-        }
-
-        [Fact]
-        public void Log_Section_logs_a_successful_result_when_completed()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit(requireConfirm: true))
-            {
-                section.Success();
-            }
-
-            log.Last().IsSectionComplete.Should().BeTrue();
-            log.Last().IsSectionSuccessful.Should().BeTrue();
-        }
-
-        [Fact]
-        public void Log_Section_logs_an_unsuccessful_result_when_not_completed()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (OnEnterAndExit(requireConfirm: true))
-            {
-            }
-
-            log.Last().IsSectionComplete.Should().BeTrue();
-            log.Last().IsSectionSuccessful.Should().BeFalse();
-        }
-
-        [Fact]
-        public void Log_Section_logs_an_unsuccessful_result_when_completed_with_Fail()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit(requireConfirm: true))
-            {
-                section.Fail();
-            }
-
-            log.Last().IsSectionComplete.Should().BeTrue();
-            log.Last().IsSectionSuccessful.Should().BeFalse();
-        }
-
-        [Fact]
-        public void Log_Section_Success_can_be_used_to_add_additional_properties_on_complete()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit())
-            {
-                section.Success("All done {0}", args: "bye!");
-            }
-
-            log.Last()
-               .Select(_ => _.Value)
-               .Should()
-               .ContainSingle(arg => arg == "bye!");
-        }
-
-        [Fact]
-        public void Log_Section_Fail_can_be_used_to_add_additional_properties_on_complete()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnEnterAndExit())
-            {
-                section.Fail(message: "Oops! {0}", args: "bye!");
-            }
-
-            log.Last()
-               .Select(_ => _.Value)
-               .Should()
-               .ContainSingle(arg => arg == "bye!");
-        }
-
-        [Fact]
-        public void After_Success_is_called_then_a_LogSection_does_not_log_again()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnExit())
-            {
-                section.Success(message: "Oops! {0}", args: "bye!");
-                section.Info("hello");
-            }
-
-            log.Count.Should().Be(1);
-        }
-
-        [Fact]
-        public void After_Fail_is_called_then_a_LogSection_does_not_log_again()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnExit())
-            {
-                section.Fail(message: "Oops! {0}", args: "bye!");
-                section.Info("hello");
-            }
-
-            log.Count.Should().Be(1);
-        }
-
-        [Fact]
-        public void After_Dispose_is_called_then_a_LogSection_does_not_log_again()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section = OnExit())
-            {
-                section.Dispose();
-                section.Info("hello");
-            }
-
-            log.Count.Should().Be(1);
-        }
-
-        [Fact]
         public void Default_logger_has_no_category()
         {
             var log = new List<LogEntry>();
@@ -438,50 +210,6 @@ namespace Pocket.Tests
             }
 
             log.Should().OnlyContain(e => e.Category == "");
-        }
-
-        [Fact]
-        public void Default_Section_has_no_category()
-        {
-            var log = new List<LogEntry>();
-
-            using (Subscribe(log.Add))
-            using (var section1 = OnExit())
-            using (var section2 = OnEnterAndExit())
-            {
-                section1.Info("hello");
-                section2.Info("hello");
-            }
-
-            log.Should().OnlyContain(e => e.Category == null);
-        }
-
-        [Fact(Skip = "not implemented yet")]
-        public void All_logs_can_be_filtered_using_middleware()
-        {
-            // TODO (Logs_can_be_enriched) write test
-            Assert.True(false, "Test Logs_can_be_enriched is not written yet.");
-        }
-
-        [Fact(Skip = "not implemented yet")]
-        public void All_logs_can_be_enriched_using_middleware()
-        {
-            // TODO (Logs_can_be_enriched) write test
-            Assert.True(false, "Test Logs_can_be_enriched is not written yet.");
-        }
-
-        [Fact(Skip = "not implemented yet")]
-        public void Exceptions_thrown_by_middleware_do_not_impact_the_caller()
-        {
-            // TODO (Exceptions_thrown_by_enrichers_do_not_impact_the_caller) write test
-            Assert.True(false, "Test Exceptions_thrown_by_enrichers_do_not_impact_the_caller is not written yet.");
-        }
-
-        [Fact(Skip = "not implemented yet")]
-        public void Loggers_in_different_assemblies_can_be_discovered_and_subscribed()
-        {
-            // TODO (Loggers_in_different_assemblies_can_be_discovered_and_subscribed) write test
-            Assert.True(false, "Test Loggers_in_different_assemblies_can_be_discovered_and_subscribed is not written yet.");
         }
     }
 }
